+++ lib/src/controls/inputs/slider.dart
@@ -1,3 +1,5 @@
+import 'dart:math' as math;
+
 import 'package:flutter/rendering.dart';
 import 'package:flutter/material.dart' as m;
 import 'package:fluent_ui/fluent_ui.dart';
@@ -50,25 +52,28 @@ class Slider extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final style = context.theme!.sliderStyle!.copyWith(this.style);
+    final style = context.theme!.sliderStyle?.copyWith(this.style);
     return Padding(
-      padding: style.margin ?? EdgeInsets.zero,
+      padding: style?.margin ?? EdgeInsets.zero,
       child: m.Material(
         type: m.MaterialType.transparency,
         child: m.SliderTheme(
           data: m.SliderThemeData(
             showValueIndicator: m.ShowValueIndicator.always,
-            thumbColor: style.thumbColor ?? style.activeColor,
+            thumbColor: style?.thumbColor ?? style?.activeColor,
             overlayShape: m.RoundSliderOverlayShape(overlayRadius: 0),
             thumbShape: m.RoundSliderThumbShape(
               elevation: 0,
               pressedElevation: 0,
             ),
+            valueIndicatorShape: _RectangularSliderValueIndicatorShape(
+              style?.labelBackgroundColor,
+            ),
             trackHeight: 0.25,
             trackShape: _CustomTrackShape(),
-            disabledThumbColor: style.disabledThumbColor,
-            disabledInactiveTrackColor: style.disabledInactiveColor,
-            disabledActiveTrackColor: style.disabledActiveColor,
+            disabledThumbColor: style?.disabledThumbColor,
+            disabledInactiveTrackColor: style?.disabledInactiveColor,
+            disabledActiveTrackColor: style?.disabledActiveColor,
           ),
           child: m.Slider(
             value: value,
@@ -77,12 +82,10 @@ class Slider extends StatelessWidget {
             onChanged: onChanged,
             onChangeEnd: onChangeEnd,
             onChangeStart: onChangeStart,
-            activeColor: style.activeColor,
-            inactiveColor: style.inactiveColor,
+            activeColor: style?.activeColor,
+            inactiveColor: style?.inactiveColor,
             divisions: divisions,
-            mouseCursor: style.cursor,
-            // TODO: improve label fidelity
-            // Image example: https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/controls/slider.png
+            mouseCursor: style?.cursor,
             label: label,
             focusNode: focusNode,
           ),
@@ -112,6 +115,7 @@ class _CustomTrackShape extends m.RoundedRectSliderTrackShape {
 class SliderStyle {
   final Color? thumbColor;
   final Color? disabledThumbColor;
+  final Color? labelBackgroundColor;
 
   final MouseCursor? cursor;
 
@@ -137,6 +141,7 @@ class SliderStyle {
     this.disabledActiveColor,
     this.inactiveColor,
     this.disabledInactiveColor,
+    this.labelBackgroundColor,
   });
 
   static SliderStyle defaultTheme(Style? style) {
@@ -144,13 +149,13 @@ class SliderStyle {
       cursor: SystemMouseCursors.click,
       thumbColor: style?.accentColor,
       activeColor: style?.accentColor,
-      inactiveColor: style?.inactiveColor,
+      inactiveColor: style?.disabledColor?.withOpacity(1),
       margin: EdgeInsets.zero,
       animationDuration: style?.animationDuration,
       animationCurve: style?.animationCurve,
-      disabledActiveColor: style?.disabledColor,
+      disabledActiveColor: style?.disabledColor?.withOpacity(1),
       disabledThumbColor: style?.disabledColor?.withOpacity(1),
-      disabledInactiveColor: style?.inactiveColor,
+      disabledInactiveColor: style?.disabledColor,
     );
 
     return def;
@@ -164,11 +169,204 @@ class SliderStyle {
       animationDuration: style?.animationDuration ?? animationDuration,
       thumbColor: style?.thumbColor ?? thumbColor,
       activeColor: style?.activeColor ?? activeColor,
-      inactiveColor: style?.inactiveColor,
+      inactiveColor: style?.inactiveColor ?? inactiveColor,
       disabledActiveColor: style?.disabledActiveColor ?? disabledActiveColor,
       disabledInactiveColor:
           style?.disabledInactiveColor ?? disabledInactiveColor,
       disabledThumbColor: style?.disabledThumbColor ?? disabledThumbColor,
+      labelBackgroundColor: style?.labelBackgroundColor ?? labelBackgroundColor,
     );
+  }
+}
+
+class _RectangularSliderValueIndicatorShape extends m.SliderComponentShape {
+  final Color? backgroundColor;
+
+  /// Create a slider value indicator that resembles a rectangular tooltip.
+  const _RectangularSliderValueIndicatorShape([this.backgroundColor]);
+
+  static const _RectangularSliderValueIndicatorPathPainter _pathPainter =
+      _RectangularSliderValueIndicatorPathPainter();
+
+  @override
+  Size getPreferredSize(
+    bool isEnabled,
+    bool isDiscrete, {
+    TextPainter? labelPainter,
+    double? textScaleFactor,
+  }) {
+    assert(labelPainter != null);
+    assert(textScaleFactor != null && textScaleFactor >= 0);
+    return _pathPainter.getPreferredSize(labelPainter!, textScaleFactor!);
+  }
+
+  @override
+  void paint(
+    PaintingContext context,
+    Offset center, {
+    required Animation<double> activationAnimation,
+    required Animation<double> enableAnimation,
+    required bool isDiscrete,
+    required TextPainter labelPainter,
+    required RenderBox parentBox,
+    required m.SliderThemeData sliderTheme,
+    required TextDirection textDirection,
+    required double value,
+    required double textScaleFactor,
+    required Size sizeWithOverflow,
+  }) {
+    final Canvas canvas = context.canvas;
+    final double scale = activationAnimation.value;
+    _pathPainter.paint(
+      parentBox: parentBox,
+      canvas: canvas,
+      center: center,
+      scale: scale,
+      labelPainter: labelPainter,
+      textScaleFactor: textScaleFactor,
+      sizeWithOverflow: sizeWithOverflow,
+      backgroundPaintColor: backgroundColor ?? sliderTheme.valueIndicatorColor!,
+    );
+  }
+}
+
+class _RectangularSliderValueIndicatorPathPainter {
+  const _RectangularSliderValueIndicatorPathPainter();
+
+  static const double _triangleHeight = 8.0;
+  static const double _labelPadding = 8.0;
+  static const double _preferredHeight = 32.0;
+  static const double _minLabelWidth = 16.0;
+  static const double _bottomTipYOffset = 14.0;
+  static const double _preferredHalfHeight = _preferredHeight / 2;
+  static const double _upperRectRadius = 4;
+
+  Size getPreferredSize(TextPainter labelPainter, double textScaleFactor) {
+    return Size(
+      _upperRectangleWidth(labelPainter, 1, textScaleFactor),
+      labelPainter.height + _labelPadding,
+    );
+  }
+
+  double getHorizontalShift({
+    required RenderBox parentBox,
+    required Offset center,
+    required TextPainter labelPainter,
+    required double textScaleFactor,
+    required Size sizeWithOverflow,
+    required double scale,
+  }) {
+    assert(!sizeWithOverflow.isEmpty);
+
+    const double edgePadding = 8.0;
+    final double rectangleWidth =
+        _upperRectangleWidth(labelPainter, scale, textScaleFactor);
+
+    /// Value indicator draws on the Overlay and by using the global Offset
+    /// we are making sure we use the bounds of the Overlay instead of the Slider.
+    final Offset globalCenter = parentBox.localToGlobal(center);
+
+    // The rectangle must be shifted towards the center so that it minimizes the
+    // chance of it rendering outside the bounds of the render box. If the shift
+    // is negative, then the lobe is shifted from right to left, and if it is
+    // positive, then the lobe is shifted from left to right.
+    final double overflowLeft =
+        math.max(0, rectangleWidth / 2 - globalCenter.dx + edgePadding);
+    final double overflowRight = math.max(
+        0,
+        rectangleWidth / 2 -
+            (sizeWithOverflow.width - globalCenter.dx - edgePadding));
+
+    if (rectangleWidth < sizeWithOverflow.width) {
+      return overflowLeft - overflowRight;
+    } else if (overflowLeft - overflowRight > 0) {
+      return overflowLeft - (edgePadding * textScaleFactor);
+    } else {
+      return -overflowRight + (edgePadding * textScaleFactor);
+    }
+  }
+
+  double _upperRectangleWidth(
+    TextPainter labelPainter,
+    double scale,
+    double textScaleFactor,
+  ) {
+    final double unscaledWidth =
+        math.max(_minLabelWidth * textScaleFactor, labelPainter.width) +
+            _labelPadding;
+    return unscaledWidth * scale;
+  }
+
+  void paint({
+    required RenderBox parentBox,
+    required Canvas canvas,
+    required Offset center,
+    required double scale,
+    required TextPainter labelPainter,
+    required double textScaleFactor,
+    required Size sizeWithOverflow,
+    required Color backgroundPaintColor,
+    Color? strokePaintColor,
+  }) {
+    if (scale == 0.0) {
+      // Zero scale essentially means "do not draw anything", so it's safe to just return.
+      return;
+    }
+    assert(!sizeWithOverflow.isEmpty);
+
+    final double rectangleWidth = _upperRectangleWidth(
+      labelPainter,
+      scale,
+      textScaleFactor,
+    );
+    final double horizontalShift = getHorizontalShift(
+      parentBox: parentBox,
+      center: center,
+      labelPainter: labelPainter,
+      textScaleFactor: textScaleFactor,
+      sizeWithOverflow: sizeWithOverflow,
+      scale: scale,
+    );
+
+    final double rectHeight = labelPainter.height + _labelPadding;
+    final Rect upperRect = Rect.fromLTWH(
+      -rectangleWidth / 2 + horizontalShift,
+      -_triangleHeight - rectHeight,
+      rectangleWidth,
+      rectHeight,
+    );
+
+    final Path trianglePath = Path()..close();
+    final Paint fillPaint = Paint()..color = backgroundPaintColor;
+    final RRect upperRRect = RRect.fromRectAndRadius(
+      upperRect,
+      const Radius.circular(_upperRectRadius),
+    );
+    trianglePath.addRRect(upperRRect);
+
+    canvas.save();
+    // Prepare the canvas for the base of the tooltip, which is relative to the
+    // center of the thumb.
+    canvas.translate(center.dx, center.dy - _bottomTipYOffset);
+    canvas.scale(scale, scale);
+    if (strokePaintColor != null) {
+      final Paint strokePaint = Paint()
+        ..color = strokePaintColor
+        ..strokeWidth = 1.0
+        ..style = PaintingStyle.stroke;
+      canvas.drawPath(trianglePath, strokePaint);
+    }
+    canvas.drawPath(trianglePath, fillPaint);
+
+    // The label text is centered within the value indicator.
+    final double bottomTipToUpperRectTranslateY =
+        -_preferredHalfHeight / 2 - upperRect.height;
+    canvas.translate(0, bottomTipToUpperRectTranslateY);
+    final Offset boxCenter = Offset(horizontalShift, upperRect.height / 2);
+    final Offset halfLabelPainterOffset =
+        Offset(labelPainter.width / 2, labelPainter.height / 2);
+    final Offset labelOffset = boxCenter - halfLabelPainterOffset;
+    labelPainter.paint(canvas, labelOffset);
+    canvas.restore();
   }
 }