+++ lib/src/utils.dart
@@ -58,7 +58,7 @@ void debugCheckHasFluentTheme(BuildContext context, [bool check = true]) {
   if (check)
     assert(
       context.theme != null,
-      'A Theme widget is necessary to draw this layout. It is implemented by default in FluentApp. ' +
+      'A Theme widget is necessary to draw this layout. It is implemented by default in FluentApp. '
           'To fix this, wrap a Theme widget upper in this layout or implement a FluentApp.',
     );
 }